Fix and Improve Looting Logic
Instead of trying to wait for a display to open, the EH simply overrides the inventory menu if you want to prevent looting.

Also added a check to see if the body's "side" is the same as the player's side. Not sure if it will actually work, because player bodies return to side "civ" when they die. Might need to tweak that later, but it gets the job done for now.

NOTE: This will allow you to open inventory if you only see a weapon on the ground, so people will still be able to take whatever weapons/attachments they want as long as they can gear up on it separately from the body.